+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add deferred key-loading function to `Decrypter`, allowing the type to be
   constructed with a closure that later provides a secret key based on the
   public key present in a stream.
+- Extend deferred key-loading to `DecryptingReader`.
 
 ### Changed
 - Change MSRV from 1.34 to 1.33